Extract addressing modes to addressing.go
