+++ cpu.go
@@ -147,134 +147,6 @@ func (c *CPU) getFlag(flag Flags) bool {
 	return (c.P & flag) > 0
 }
 
-// --- Addressing Modes ---
-// All addressing mode functions remain the same method signatures:
-// func (c *CPU) ModeName() uint8 { ... }
-
-func (c *CPU) IMP() uint8 {
-	return 0
-}
-
-func (c *CPU) IMM() uint8 {
-	c.addrAbs = c.PC
-	c.PC++
-	return 0
-}
-
-func (c *CPU) ZP0() uint8 {
-	c.addrAbs = uint16(c.read(c.PC))
-	c.PC++
-	c.addrAbs &= 0x00FF
-	return 0
-}
-
-func (c *CPU) ZPX() uint8 {
-	baseAddr := uint16(c.read(c.PC))
-	c.PC++
-	c.addrAbs = (baseAddr + uint16(c.X)) & 0x00FF
-	return 0
-}
-
-func (c *CPU) ZPY() uint8 {
-	baseAddr := uint16(c.read(c.PC))
-	c.PC++
-	c.addrAbs = (baseAddr + uint16(c.Y)) & 0x00FF
-	return 0
-}
-
-func (c *CPU) REL() uint8 {
-	addrRelOffset := uint16(c.read(c.PC))
-	c.PC++
-	if addrRelOffset&0x80 > 0 {
-		addrRelOffset |= 0xFF00
-	}
-	c.addrRel = c.PC + addrRelOffset
-	return 0
-}
-
-func (c *CPU) ABS() uint8 {
-	lo := uint16(c.read(c.PC))
-	c.PC++
-	hi := uint16(c.read(c.PC))
-	c.PC++
-	c.addrAbs = (hi << 8) | lo
-	return 0
-}
-
-func (c *CPU) ABX() uint8 {
-	lo := uint16(c.read(c.PC))
-	c.PC++
-	hi := uint16(c.read(c.PC))
-	c.PC++
-	baseAddr := (hi << 8) | lo
-	c.addrAbs = baseAddr + uint16(c.X)
-
-	if (c.addrAbs & 0xFF00) != (baseAddr & 0xFF00) {
-		return 1
-	}
-	return 0
-}
-
-func (c *CPU) ABY() uint8 {
-	lo := uint16(c.read(c.PC))
-	c.PC++
-	hi := uint16(c.read(c.PC))
-	c.PC++
-	baseAddr := (hi << 8) | lo
-	c.addrAbs = baseAddr + uint16(c.Y)
-
-	if (c.addrAbs & 0xFF00) != (baseAddr & 0xFF00) {
-		return 1
-	}
-	return 0
-}
-
-func (c *CPU) IND() uint8 {
-	ptrLo := uint16(c.read(c.PC))
-	c.PC++
-	ptrHi := uint16(c.read(c.PC))
-	c.PC++
-	ptr := (ptrHi << 8) | ptrLo
-
-	if c.variant.HasIndirectJMPBug() && ptrLo == 0x00FF {
-		// NMOS bug: If low byte is $FF, high byte is fetched from $xx00
-		c.addrAbs = uint16(c.read(ptr)) | (uint16(c.read(ptr&0xFF00)) << 8)
-	} else {
-		// CMOS fix: Normal behavior
-		c.addrAbs = (uint16(c.read(ptr+1)) << 8) | uint16(c.read(ptr))
-	}
-	return 0
-}
-
-func (c *CPU) IZX() uint8 {
-	zpAddrBase := uint16(c.read(c.PC))
-	c.PC++
-	// Add X and wrap around zero page
-	zpAddr := (zpAddrBase + uint16(c.X)) & 0x00FF
-
-	// Read the effective address from zero page, wrapping around if needed
-	effAddrLo := uint16(c.read(zpAddr))
-	effAddrHi := uint16(c.read((zpAddr + 1) & 0x00FF))
-	c.addrAbs = (effAddrHi << 8) | effAddrLo
-	return 0
-}
-
-func (c *CPU) IZY() uint8 {
-	zpAddr := uint16(c.read(c.PC))
-	c.PC++
-
-	baseAddrLo := uint16(c.read(zpAddr & 0x00FF))
-	baseAddrHi := uint16(c.read((zpAddr + 1) & 0x00FF))
-	baseAddr := (baseAddrHi << 8) | baseAddrLo
-
-	c.addrAbs = baseAddr + uint16(c.Y)
-
-	if (c.addrAbs & 0xFF00) != (baseAddr & 0xFF00) {
-		return 1
-	}
-	return 0
-}
-
 // --- Lookup Table Initialization ---
 // *Use method expressions for assignment*
 func (c *CPU) buildLookupTable() {
